src/python/ml_downloader/requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
MySQL-python==1.2.5
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
SQLAlchemy-Utils==0.32.21
urllib3==1.22
